Apply custom column width only in advanced view
The chaos.social default theme widens the columns in the multi-column (advanced) view. However, the maximum width also applied to the new single-column (standard) view, so that toots didn't use all available space.

This change limits the width change to the multi-column view, allowing the timeline to take up the full width in the new standard view, too.